+++ index.js
@@ -230,13 +230,35 @@ module.exports = {
 				'**/?(*.)+(spec|test).[jt]s?(x)',
 				'**/*-specs.{js,jsx,ts,tsx}'
 			],
+			plugins: ['jest', 'testing-library'],
 			excludedFiles: ['tests/screenshot/**/*', 'tests/ui/**/*'],
 			env: {
-				jest: true
+				jest: true,
+				'jest/globals': true,
 			},
 			rules: {
 				// Arrow functions can simplify tests
-				'react/jsx-no-bind': 'off'
+				'react/jsx-no-bind': 'off',
+
+				// https://github.com/jest-community/eslint-plugin-jest
+				'jest/no-conditional-expect': 'error',
+				'jest/no-identical-title': 'error',
+				'jest/no-interpolation-in-snapshots': 'error',
+				'jest/no-jasmine-globals': 'error',
+				'jest/no-jest-import': 'error',
+				'jest/no-mocks-import': 'error',
+				'jest/valid-describe': 'error',
+				'jest/valid-expect': 'error',
+				'jest/valid-expect-in-promise': 'error',
+				'jest/valid-title': 'warn',
+
+				// https://github.com/testing-library/eslint-plugin-testing-library
+				'testing-library/await-async-query': 'error',
+				'testing-library/await-async-utils': 'error',
+				'testing-library/no-await-sync-query': 'warn',
+				'testing-library/no-dom-import': ['error', 'react'],
+				'testing-library/no-wait-for-empty-callback': 'error',
+				'testing-library/no-wait-for-snapshot': 'error',
 			}
 		},
 		{